scripts/requirements.txt
PyYAML~=6.0.1
pillow~=10.2.0